Rename to skewed_allocator.hpp.
